pyproject.toml
[tool.poetry]
name = "garagem2.0"
version = "0.1.0"
description = ""
authors = ["zCrazyy <carlos.fraga2004@gmail.com>"]

[tool.poetry.dependencies]
python = "^3.8"
Django = "^4.1"
djangorestframework = "^3.13.1"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
